requirements.txt
flask
line-bot-sdk
gspread
oauth2client
python-dotenv
rich
pillow
gunicorn
google-api-python-client
google-auth
requests